+++ src/sagemaker/sklearn/estimator.py
@@ -75,7 +75,7 @@ class SKLearn(Framework):
             framework_version (str): Scikit-learn version you want to use for
                 executing your model training code. Defaults to ``None``. Required
                 unless ``image_uri`` is provided. List of supported versions:
-                https://github.com/aws/sagemaker-python-sdk#sklearn-sagemaker-estimators
+                https://docs.aws.amazon.com/sagemaker/latest/dg/sklearn.html
             py_version (str): Python version you want to use for executing your
                 model training code (default: 'py3'). Currently, 'py3' is the only
                 supported version. If ``None`` is passed in, ``image_uri`` must be